support shuffle on photo albums
