+++ MediaBrowser.Model/System/IPowerManagement.cs
@@ -1,9 +1,11 @@
-﻿
+﻿using System;
+
 namespace MediaBrowser.Model.System
 {
     public interface IPowerManagement
     {
         void PreventSystemStandby();
         void AllowSystemStandby();
+        void ScheduleWake(DateTime wakeTimeUtc, string displayName);
     }
 }
